Implement command line app for testing purposes
This is a minimal command line application that allows to send supported SCPI (and RAW) commands to instrument and provides logging of the received data.

When logging is enabled the received data is entirely logged to the file! This means that for example DISPLAY_DATA command will log TMC block header to log file as well!